Move the schema.org scripts into another module because they're verbose and distracting
