+++ src/templates/layouts/vetCenter/index.tsx
@@ -10,134 +10,40 @@ import { QaSection } from '@/templates/components/qaSection'
 import { Accordion } from '@/templates/components/accordion'
 import { ExpandableOperatingStatus } from './ExpandableOperatingStatus'
 import { PhoneNumber } from '@/templates/common/phoneNumber'
+import { SchemaScript } from './SchemaScript'
 
-export function VetCenter({
-  address,
-  ccNonTraditionalHours,
-  ccVetCenterCallCenter,
-  ccVetCenterFaqs,
-  geolocation,
-  featuredContent,
-  introText,
-  missionExplainer,
-  officeHours,
-  officialName,
-  operatingStatusFacility,
-  operatingStatusMoreInfo,
-  phoneNumber,
-  healthServices,
-  counselingHealthServices,
-  referralHealthServices,
-  otherHealthServices,
-  image,
-  bannerImage,
-  prepareForVisit,
-  title,
-  fieldFacilityLocatorApiId,
-  path,
-}: FormattedVetCenter) {
+export function VetCenter(vetCenterProps: FormattedVetCenter) {
+  const {
+    address,
+    ccNonTraditionalHours,
+    ccVetCenterCallCenter,
+    ccVetCenterFaqs,
+    geolocation,
+    featuredContent,
+    introText,
+    missionExplainer,
+    officeHours,
+    officialName,
+    operatingStatusFacility,
+    operatingStatusMoreInfo,
+    phoneNumber,
+    healthServices,
+    counselingHealthServices,
+    referralHealthServices,
+    otherHealthServices,
+    image,
+    bannerImage,
+    prepareForVisit,
+    title,
+    fieldFacilityLocatorApiId,
+    path,
+  } = vetCenterProps
+
   const directionsString = [
     address?.address_line1,
     address?.locality,
     address?.administrative_area,
   ]
-
-  const structuredSchemaData = {
-    '@context': 'https://schema.org',
-    '@type': 'Place',
-    address: {
-      '@type': 'PostalAddress',
-      streetAddress: `${address.address_line1}${
-        address.address_line2 ? `, ${address.address_line2}` : ''
-      }`,
-      addressLocality: address.locality,
-      addressRegion: address.administrative_area,
-      postalCode: address.postal_code,
-    },
-    name: title,
-    telephone: phoneNumber,
-    openingHoursSpecification: officeHours.map((hours) => ({
-      '@type': 'OpeningHoursSpecification',
-      dayOfWeek: `https://schema.org/${hours.day}`,
-      opens: hours.starthours,
-      closes: hours.endhours,
-    })),
-    hasMap: `https://maps.google.com?saddr=Current+Location&daddr=${encodeURIComponent(
-      `${address.address_line1}, ${address.locality}, ${address.postal_code}`
-    )}`,
-    image: image?.links?.['2_1_large']?.href,
-    branchCode: fieldFacilityLocatorApiId,
-    geo: {
-      '@type': 'GeoCoordinates',
-      latitude: geolocation.lat.toString(),
-      longitude: geolocation.lon.toString(),
-    },
-  }
-
-  const generateStructuredDataForHealthServices = (
-    healthServices,
-    title,
-    fieldAddress
-  ) => {
-    return healthServices.map((service) => ({
-      '@context': 'https://schema.org',
-      '@type': 'GovernmentService',
-      name: title,
-      alternateName: service?.fieldVetCenterFriendlyName || null,
-      serviceType: service?.vetCenterTypeOfCare || null,
-      serviceOperator: {
-        '@type': 'GovernmentOrganization',
-        name: 'US Department of Veterans Affairs',
-      },
-      areaServed: {
-        '@type': 'AdministrativeArea',
-        name: fieldAddress?.administrative_area || null,
-      },
-      audience: {
-        '@type': 'Audience',
-        audienceType: 'Veteran',
-      },
-      availableChannel: {
-        '@type': 'ServiceChannel',
-        serviceUrl: 'https://www.va.gov',
-        servicePhone: {
-          '@type': 'ContactPoint',
-          telephone: service?.phoneNumber || null,
-        },
-      },
-      provider: {
-        '@type': 'GovernmentOrganization',
-        name: 'Veterans Affairs',
-        url: 'https://www.va.gov',
-      },
-      serviceLocation: {
-        '@type': 'Place',
-        name: title,
-        address: {
-          '@type': 'PostalAddress',
-          streetAddress:
-            fieldAddress.address_line1 +
-            (address.address_line2
-              ? `${address.address_line1}${address.address_line2}`
-              : address.address_line1),
-          addressLocality: fieldAddress?.locality,
-          addressRegion: fieldAddress?.administrative_area,
-          postalCode: fieldAddress?.postal_code,
-        },
-        geo: {
-          '@type': 'GeoCoordinates',
-          latitude: geolocation.lat.toString(),
-          longitude: geolocation.lon.toString(),
-        },
-      },
-    }))
-  }
-
-  const structuredDataHealthServices = generateStructuredDataForHealthServices(
-    healthServices,
-    title,
-    address
-  )
 
   const PrepareForVisitComponent = ({ visitItems }) => {
     if (visitItems.length === 0) return null
@@ -346,20 +252,7 @@ export function VetCenter({
 
           <va-back-to-top></va-back-to-top>
 
-          {/* Embedding structured data scripts for schema.org */}
-          <script
-            type="application/ld+json"
-            dangerouslySetInnerHTML={{
-              __html: JSON.stringify(structuredSchemaData),
-            }}
-          />
-          {structuredDataHealthServices.map((service, index) => (
-            <script
-              key={index}
-              type="application/ld+json"
-              dangerouslySetInnerHTML={{ __html: JSON.stringify(service) }}
-            />
-          ))}
+          <SchemaScript vetCenter={vetCenterProps} />
         </article>
       </div>
     </div>
